requirements.txt
fastapi
uvicorn
python-multipart
pillow
pymongo
minio
